requirements.txt
requests
pandas
jupyter
ipykernel
python-dotenv
